+++ src/skyperious/gui.py
@@ -8,7 +8,7 @@ Released under the MIT License.
 
 @author      Erki Suurjaak
 @created     26.11.2011
-@modified    27.04.2022
+@modified    29.04.2022
 ------------------------------------------------------------------------------
 """
 import ast
@@ -3603,12 +3603,7 @@ class DatabasePage(wx.Panel):
                 else:
                     plabel = u"Synchronizing %s" % result["table"]
                     if "messages" == result["table"] and result.get("chat"):
-                        chat = next((c for c in self.chats if c["identity"] == result["chat"]), None)
-                        if not chat:
-                            cc = self.db.get_conversations(chatidentities=[result["chat"]], reload=True, log=False)
-                            self.chats.extend(cc)
-                            if cc: chat = cc[0]
-
+                        chat  = next(c for c in self.chats if c["identity"] == result["chat"])
                         title = chat["title_long_lc"] if chat else result["chat"]
                         if len(title) > 35:
                             title = title[:35] + ".."
@@ -3641,9 +3636,12 @@ class DatabasePage(wx.Panel):
                                 ]))
 
                             self.chats = self.db.get_conversations(reload=True, log=False)
+
                             def after2():
-                                if self: self.db.get_conversations_stats(self.chats)
-                                if self: self.list_chats.Populate(self.chats)
+                                """Populate statistics for all chats after completion."""
+                                if not self: return
+                                self.db.get_conversations_stats(self.chats)
+                                self.list_chats.Populate(self.chats)
                             wx.CallAfter(after2)
 
                         if "messages" == result["table"]:
@@ -3657,9 +3655,9 @@ class DatabasePage(wx.Panel):
                                 self.list_chats_sync.AppendRow(row)
                                 self.list_chats_sync.ResetColumnWidths()
                                 chat["message_count" ] = (chat["message_count"] or 0) + result["new"]
-                                chat["first_message_datetime"]  = min(chat["first_message_datetime"] or result["first"], result["first"])
-                                chat["last_message_datetime"]   = max(chat["last_message_datetime"]  or result["first"], result["last"])
-                                chat["last_activity_datetime"]  = max(chat["last_activity_datetime"] or result["last"],  result["last"])
+                                chat["first_message_datetime"] = min(chat["first_message_datetime"] or result["first"], result["first"])
+                                chat["last_message_datetime"]  = max(chat["last_message_datetime"]  or result["first"], result["last"])
+                                chat["last_activity_datetime"] = max(chat["last_activity_datetime"] or result["last"],  result["last"])
                                 for k in "first_message", "last_message", "last_activity":
                                     chat[k + "_timestamp"] = util.datetime_to_epoch(chat[k + "_datetime"])
                                 self.list_chats.Populate(self.chats)
@@ -3670,6 +3668,36 @@ class DatabasePage(wx.Panel):
                                 slabel += ": %s new" % result["new"]
                                 if result["updated"]: slabel += ", %s updated" % result["updated"]
                                 slabel += "."
+
+                            def after2():
+                                """Populate statistics for chat participants after finishing chat messages."""
+                                ids = [x["identity"] for x in chat["participants"]]
+                                contacts = [x for x in self.contacts if x["identity"] in ids]
+                                self.db.get_contacts_stats(contacts, self.chats)
+                                self.list_contacts.Populate(self.contacts)
+                                if self.contact and self.contact["identity"] in ids:
+                                    contact = next((x for x in self.contacts
+                                                    if x["identity"] == self.contact["identity"]), None)
+                                    if contact: self.load_contact(contact)
+                            chat and wx.CallAfter(after2)
+
+                    elif not result.get("start"):
+                        if "chats" == result["table"] and result.get("chat"):
+                            # Inserted or updated a chat: load new contacts, if any
+                            plabel = ""
+
+                            chat = next((c for c in self.chats if c["identity"] == result["chat"]), None)
+                            if not chat:
+                                cc = self.db.get_conversations(chatidentities=[result["chat"]], reload=True, log=False)
+                                self.chats.extend(cc)
+                                if cc: chat = cc[0]
+                            if chat:
+                                ids = [x["identity"] for x in chat["participants"]]
+                                for identity in set(ids) - set(x["identity"] for x in self.contacts):
+                                    c = self.db.get_contact(identity)
+                                    if c:
+                                        self.contacts.append(c)
+                                        self.list_contacts.Populate(self.contacts)
 
                 if plabel:
                     self.label_sync_progress.Label = plabel