+++ CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 
 5.0, 2022-04-27
 ---------------
+- refresh contacts list during and after live sync;
 - fixed error on filtering current chat by text (#107).